Got thunder tests working
