+++ flask-app/thunder-tests/thunderCollection.json
@@ -1,0 +1,10 @@
+[
+  {
+    "_id": "9ba952ff-523c-4c1d-a433-f84198139574",
+    "colName": "test",
+    "created": "2023-04-17T02:56:01.145Z",
+    "sortNum": 10000,
+    "folders": [],
+    "settings": {}
+  }
+]
