+++ flask-app/thunder-tests/thunderclient.json
@@ -1,0 +1,1 @@
+[]
